NEW: Rust Api & C Api are now more isolated NEW: `toncli` is now compilable
