+++ Cargo.toml
@@ -2,7 +2,7 @@
 members = [
     "ton_sdk",
     "ton_client/client",
-    "ton_sdk_cli",
+    "toncli",
     "api/test",
 ]
