+++ ton_client/client/src/api/api_reference.rs
@@ -1,22 +1,5 @@
-use crate::api::get_dispatcher;
-use crate::client::ClientContext;
-use crate::error::ApiResult;
 use api_info::{Function, Type, API};
 use std::collections::HashMap;
-use std::sync::Arc;
-
-#[derive(ApiType, Serialize, Deserialize)]
-pub struct ResultOfGetApiReference {
-    pub api: API,
-}
-
-#[api_function]
-pub(crate) fn get_api_reference(_context: Arc<ClientContext>) -> ApiResult<ResultOfGetApiReference> {
-    let api = ApiReducer::build(&get_dispatcher().api);
-    Ok(ResultOfGetApiReference {
-        api,
-    })
-}
 
 fn is_full_name(name: &str) -> bool {
     name.contains(".")
