+++ ton_client/client/src/client/mod.rs
@@ -19,15 +19,17 @@ mod std_client_env;
 mod tests;
 
 pub use client::{
-    Client, ClientConfig, ClientContext, ContextHandle, CryptoConfig,
+    ClientConfig, ClientContext, CryptoConfig,
 };
 pub use errors::{Error, ErrorCode};
 
-pub(crate) use client::parse_params;
 pub(crate) use client_env::{ClientEnv, FetchMethod, FetchResult, WebSocket};
 
 use crate::error::ApiResult;
 use std::sync::Arc;
+use api_info::API;
+use crate::api::api_reference::ApiReducer;
+use crate::api::runtime::{Runtime};
 
 #[derive(Serialize, Deserialize, ApiType, Clone)]
 pub struct ResultOfVersion {
@@ -41,3 +43,17 @@ pub fn version(_context: Arc<ClientContext>) -> ApiResult<ResultOfVersion> {
         version: env!("CARGO_PKG_VERSION").to_owned(),
     })
 }
+
+#[derive(ApiType, Serialize, Deserialize)]
+pub struct ResultOfGetApiReference {
+    pub api: API,
+}
+
+#[api_function]
+pub fn get_api_reference(_context: Arc<ClientContext>) -> ApiResult<ResultOfGetApiReference> {
+    let api = ApiReducer::build(Runtime::api());
+    Ok(ResultOfGetApiReference {
+        api,
+    })
+}
+
