+++ toncli/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "tonsdk"
+name = "toncli"
 version = "0.26.0"
-description = "Command line tool for TON SDK"
+description = "TON CLient Command Line Tool"
 authors = ["TON DEV SOLUTIONS LTD <support@tonlabs.io>"]
 repository = "https://github.com/tonlabs/TON-SDK"
 homepage = "https://ton.dev"
@@ -15,10 +15,13 @@ edition = "2018"
 json5 = "0.2.8"
 regex = "1.3.9"
 dirs = "3.0.1"
-serde_json = "1.0.57"
+serde_json = "1.0.58"
+serde_derive = "1.0.116"
+serde = "1.0.116"
 
 ton_client = { path = '../ton_client/client',  default_features = false, features = ["node_interaction"] }
+api_info = { path = '../api/info' }
 
 [dev-dependencies]
-assert_cmd = "0.11"
-predicates = "1"
+assert_cmd = "1.0.1"
+predicates = "1.0.5"
