+++ toncli/tests/cli.rs
